messages/requirements.txt
confluent_kafka
anyio==3.6.2
async-timeout==4.0.2
fastapi==0.92.0
idna==3.4
kafka-python==2.0.2
pydantic==1.10.5
python-dotenv==0.21.1
sniffio==1.3.0
starlette==0.25.0
typing-extensions==4.5.0
urllib3==1.26.9
uvicorn==0.20.0